Added vanilla default slash menu items
