+++ packages/core/src/index.ts
@@ -10,6 +10,7 @@ export * from "./editor/selectionTypes";
 export * from "./extensions/FormattingToolbar/FormattingToolbarPlugin";
 export * from "./extensions/HyperlinkToolbar/HyperlinkToolbarPlugin";
 export * from "./extensions/ImageToolbar/ImageToolbarPlugin";
+export * from "./extensions/SuggestionMenu/defaultSlashMenuItems";
 export * from "./extensions/SuggestionMenu/SuggestionPlugin";
 export * from "./extensions/SideMenu/SideMenuPlugin";
 export * from "./extensions/TableHandles/TableHandlesPlugin";
